requirements.txt
vosk~=0.3.32
sounddevice~=0.4.4
matplotlib~=3.5.2
fuzzywuzzy~=0.18.0
num2words~=0.5.10
torch~=1.10.0
opencv-python~=4.5.5.64
praw~=7.6.0
requests~=2.27.1
numpy~=1.22.3
Pillow~=9.1.1
psutil~=5.9.0
selenium~=4.1.5